requirements.txt
pygame
rembg
Pillow
numpy
opencv-python
